requirements.txt
# List all third-party packages to add to your local environment

ipykernel
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
numpy
pandas
matplotlib
nltk
transformers
torch
torchvision
torchaudio
pytest
colorama
datetime
python-dotenv